Tidied some mess up

+++ src/main/Stepper.cpp
@@ -134,7 +134,7 @@ void Stepper::resetStepperMovement() {
 }
 
 // Prepare the next motor-movement of the selected motor (reset info from previous movement, set new requested number of steps)
-void Stepper::prepareMovement(int angleRequested, ISR_Flags *flags) {
+void Stepper::prepareMovement(double angleRequested, ISR_Flags *flags) {
     unsigned long stepsRequested = angle2Steps(angleRequested);
     _dir = stepsRequested > 0 ? 0 : 1;
     _totalStepsRequested = abs(stepsRequested);
